FIXED OP: 8xy5, 8xy6, 8xy7, 8xyE to match Tests

+++ CPU.py
@@ -22,7 +22,7 @@ class CPU(object):
         
         # PC
         self.programCounter = 0x200;
-        self.indexRegister = 0;
+        self.indexRegister = numpy.uint16(0);
         
         # TIMERS
         self.delayTimer = 0;
